+++ src/Bundle/AutoMapperBundle/Tests/Fixtures/Order.php
@@ -1,0 +1,14 @@
+<?php
+
+namespace Jane\Bundle\AutoMapperBundle\Tests\Fixtures;
+
+use Money\Money;
+
+class Order
+{
+    /** @var int */
+    public $id;
+
+    /** @var Money */
+    public $price;
+}