+++ CHANGELOG.md
@@ -6,6 +6,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- [AutoMapper] [GH#462](https://github.com/janephp/janephp/pull/462) Move bundle out of the component
+
 ### Changed
 - [Jane] [GH#464](https://github.com/janephp/janephp/pull/464) Update version deps for 7-dev
 - [Jane] [GH#457](https://github.com/janephp/janephp/pull/457) **BC-BREAK** Move components into `component` folder and update namespaces